Add files via upload
through this project I have tried to demonstrate my skill set to generate a report for a company with respect to its employees in the last 10 years time period. i have used SQL to solve the general queries with respect to the dataset and created a fantastic dashboard using PowerBI tool.